refactor(tui): resolve code review findings for screen colocate refactor
